Update product and inventory models to support decimal quantities: Change stock_quantity and related fields to DecimalField for improved precision. Enhance product unit handling in the frontend, allowing fractional quantities for specific units. Update serializers and migrations accordingly to reflect these changes.

+++ popy-be/apps/catalog/models.py
@@ -31,7 +31,7 @@ class Product(TimeStampedModel):
     cost_price = models.DecimalField(max_digits=12, decimal_places=2)
     selling_price = models.DecimalField(max_digits=12, decimal_places=2)
     reorder_level = models.IntegerField(default=0)
-    stock_quantity = models.IntegerField(default=0)
+    stock_quantity = models.DecimalField(max_digits=14, decimal_places=3, default=0)
     status = models.BooleanField(default=True)
 
     class Meta:
